refactor: message 디렉터리 분리 및 Cookie 이동
