+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -11,9 +11,9 @@ import nextstep.jwp.model.User;
 import org.apache.coyote.Processor;
 import org.apache.coyote.http11.exception.UnsupportedContentTypeException;
 import org.apache.coyote.http11.message.HttpMethod;
-import org.apache.coyote.http11.message.HttpRequest;
-import org.apache.coyote.http11.message.HttpResponse;
 import org.apache.coyote.http11.message.HttpStatus;
+import org.apache.coyote.http11.message.request.HttpRequest;
+import org.apache.coyote.http11.message.response.HttpResponse;
 import org.apache.coyote.http11.session.Session;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
