+++ tomcat/src/test/java/org/apache/coyote/http11/message/ContentTypeTest.java
@@ -9,6 +9,7 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import org.apache.coyote.http11.exception.UnsupportedContentTypeException;
+import org.apache.coyote.http11.message.request.HttpRequest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
